Adds CI/CD on Github Actions and Ubuntu packaging
This commit adds CI/CD on Githbub Actions with Debian/Ubuntu packaging
enabled.

Github Actions:
The github action config can be found in .github/workflows/cicd.yml.

Currently only the Debian/Ubuntu package is built and uploaded to a
personal PPA, but others (RPMs, arm builds, MacOS, Windows, Docker) are
templated and will be tackled in the future.

Building will only happen on the "main" branch. This means that we'll
start using feature branches more often.

For the Debian/Ubuntu package, we uppload to
wouterdebie/wouterdebie.github.io and generate all files necessary to
get apt-get to install packages. Note that in order to make this work,
    we need an armored GPG priv key set as the PRIV_KEY github secret
    and a personal token to be able to commit to
    wouterdebie/wouterdebie.github.io as the API_KEY github secret.