Refresh stream URL after 2:45h and fix in path handling
This commit enables the refresh of the stream URL after 2:45h of
streaming. After about 3 hours, locast.org empties the initial m3u file,
even though the file is still available. We do this by counting down
from 9900 seconds and when the counter < 0, we download the m3u URL for
the stream again and continue.

Next to that this commit also fixes an issue with paths, which caused
email confirmation screens not to be passed through.

Both issues are described in #34 and should be fixed by this commit.