Fixes syslog, Debian/Ubuntu log using --logfile
This relates to #12. Syslog logger doesn't rely on a path, but relies on
the syslog() syscall. In order to give the user some freedom on where to
log, Ubuntu/Debian packages start with --logfile
/var/log/locast2tuner/locast2tuner.log by default. The systemd service
file has been altered to stat with this option and the postinst script
creates the directory and sets the right permissions.