Fix retry behavior when streaming. Fixes #43.
Instead of killing the stream when the m3u playlist can't be fetched
from locast.org, we just warn errors and continue to serve the segments
we know of. This helps with transient errors. The stream will end if
there are no more known and unplayed segments left, but we'll try to
grab new segments every stream iteration.